Creando clase BaseParams
