+++ frontend/server/src/ContestParams.php
@@ -2,7 +2,7 @@
 
 namespace OmegaUp;
 
-class ContestParams {
+class ContestParams extends BaseParams {
     // Constants for admission mode.
     const CONTEST_ADMISSION_MODE_PRIVATE = 'private';
     const CONTEST_ADMISSION_MODE_REGISTRATION = 'registration';
@@ -75,39 +75,33 @@ class ContestParams {
 
     /**
      * @readonly
-     * @var string
+     * @var null|string
      */
     public $contestAlias;
 
     /**
      * @readonly
-     * @var string
+     * @var null|string
      */
     public $title;
 
     /**
      * @readonly
-     * @var string
+     * @var null|string
      */
     public $description;
 
     /**
      * @readonly
-     * @var int
+     * @var \OmegaUp\Timestamp
      */
     public $startTime;
 
     /**
      * @readonly
-     * @var int
+     * @var \OmegaUp\Timestamp
      */
     public $finishTime;
-
-    /**
-     * @readonly
-     * @var int
-     */
-    public $lastUpadted;
 
     /**
      * @readonly
@@ -117,61 +111,55 @@ class ContestParams {
 
     /**
      * @readonly
-     * @var int|null
-     */
-    public $rerunId;
-
-    /**
-     * @readonly
      * @var \OmegaUp\ContestParams::CONTEST_ADMISSION_MODE_PRIVATE|\OmegaUp\ContestParams::CONTEST_ADMISSION_MODE_REGISTRATION|\OmegaUp\ContestParams::CONTEST_ADMISSION_MODE_PUBLIC
      */
     public $admissionMode;
 
     /**
      * @readonly
-     * @var int
+     * @var float|null
      */
     public $scoreboard;
 
     /**
      * @readonly
-     * @var float
+     * @var float|null
      */
     public $pointsDecayFactor;
 
     /**
      * @readonly
-     * @var int
+     * @var int|null
      */
     public $submissionsGap;
 
     /**
      * @readonly
-     * @var \OmegaUp\ContestParams::CONTEST_FEEDBACK_NONE|\OmegaUp\ContestParams::CONTEST_FEEDBACK_SUMMARY|\OmegaUp\ContestParams::CONTEST_FEEDBACK_DETAILED
+     * @var \OmegaUp\ContestParams::CONTEST_FEEDBACK_NONE|\OmegaUp\ContestParams::CONTEST_FEEDBACK_SUMMARY|\OmegaUp\ContestParams::CONTEST_FEEDBACK_DETAILED|null
      */
     public $feedback;
 
     /**
      * @readonly
-     * @var int
+     * @var int|null
      */
     public $penalty;
 
     /**
      * @readonly
-     * @var \OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_CONTEST_START|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_PROBLEM_OPEN|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_RUNTIME|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_NONE
+     * @var \OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_CONTEST_START|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_PROBLEM_OPEN|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_RUNTIME|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_NONE|null
      */
     public $penaltyType;
 
     /**
      * @readonly
-     * @var \OmegaUp\ContestParams::CONTEST_PENALTY_CALC_POLICY_SUM|\OmegaUp\ContestParams::CONTEST_PENALTY_CALC_POLICY_MAX
+     * @var \OmegaUp\ContestParams::CONTEST_PENALTY_CALC_POLICY_SUM|\OmegaUp\ContestParams::CONTEST_PENALTY_CALC_POLICY_MAX|null
      */
     public $penaltyCalcPolicy;
 
     /**
      * @readonly
-     * @var int
+     * @var bool|null
      */
     public $showScoreboardAfter;
 
@@ -183,31 +171,13 @@ class ContestParams {
 
     /**
      * @readonly
-     * @var bool
-     */
-    public $urgent;
-
-    /**
-     * @readonly
-     * @var bool
-     */
-    public $recommended;
-
-    /**
-     * @readonly
-     * @var bool
-     */
-    public $archived;
-
-    /**
-     * @readonly
      * @var int|null
      */
     public $certificateCutoff;
 
     /**
      * @readonly
-     * @var \OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_UNINITIATED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_QUEUED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_GENERATED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_RETRYABLE_ERROR|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_FATAL_ERROR
+     * @var \OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_UNINITIATED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_QUEUED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_GENERATED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_RETRYABLE_ERROR|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_FATAL_ERROR|null
      */
     public $certificatesStatus;
 
@@ -225,15 +195,9 @@ class ContestParams {
 
     /**
      * @readonly
-     * @var \OmegaUp\ContestParams::CONTEST_SCORE_MODE_PARTIAL|\OmegaUp\ContestParams::CONTEST_SCORE_MODE_ALL_OR_NOTHING|\OmegaUp\ContestParams::CONTEST_SCORE_MODE_MAX_PER_GROUP
+     * @var \OmegaUp\ContestParams::CONTEST_SCORE_MODE_PARTIAL|\OmegaUp\ContestParams::CONTEST_SCORE_MODE_ALL_OR_NOTHING|\OmegaUp\ContestParams::CONTEST_SCORE_MODE_MAX_PER_GROUP|null
      */
     public $scoreMode;
-
-    /**
-     * @readonly
-     * @var bool
-     */
-    public $plagiarismTreshold;
 
     /**
      * @readonly
@@ -242,17 +206,30 @@ class ContestParams {
     public $checkPlagiarism;
 
     /**
-     * @param array{alias: string, title: string, description: string, start_time: int, finish_time: int, last_updated: int, window_length?: int, rerun_id?: int, admission_mode: \OmegaUp\ContestParams::CONTEST_ADMISSION_MODE_PRIVATE|\OmegaUp\ContestParams::CONTEST_ADMISSION_MODE_REGISTRATION|\OmegaUp\ContestParams::CONTEST_ADMISSION_MODE_PUBLIC, scoreboard: int, points_decay_factor: float, submissions_gap: int, feedback: \OmegaUp\ContestParams::CONTEST_FEEDBACK_NONE|\OmegaUp\ContestParams::CONTEST_FEEDBACK_SUMMARY|\OmegaUp\ContestParams::CONTEST_FEEDBACK_DETAILED, penalty: int, penalty_type: \OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_CONTEST_START|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_PROBLEM_OPEN|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_RUNTIME|\OmegaUp\ContestParams::CONTEST_PENALTY_TYPE_NONE, penalty_calc_policy: \OmegaUp\ContestParams::CONTEST_PENALTY_CALC_POLICY_SUM|\OmegaUp\ContestParams::CONTEST_PENALTY_CALC_POLICY_MAX, show_scoreboard_after: int, languages?: list<string>, urgent: bool, recommended: bool, archived: bool, certificate_cutoff?: int, certificates_status: \OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_UNINITIATED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_QUEUED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_GENERATED|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_RETRYABLE_ERROR|\OmegaUp\ContestParams::CONTEST_CERTIFICATE_STATUS_FATAL_ERROR, contest_for_teams?: bool, default_show_all_contestants_in_scoreboard?: bool, score_mode: \OmegaUp\ContestParams::CONTEST_SCORE_MODE_PARTIAL|\OmegaUp\ContestParams::CONTEST_SCORE_MODE_ALL_OR_NOTHING|\OmegaUp\ContestParams::CONTEST_SCORE_MODE_MAX_PER_GROUP, plagiarism_treshold: bool, check_plagiarism: bool} $params
+     * @psalm-suppress InvalidArrayOffset
+     * @param array{admission_mode: "private"|"public"|"registration", alias: null|string, check_plagiarism: bool, contest_for_teams: bool, description: null|string, feedback: "detailed"|"none"|"summary"|null, finish_time: \OmegaUp\Timestamp, languages: null|string, penalty: int|null, penalty_calc_policy: "max"|"sum"|null, penalty_type: "contest_start"|"none"|"problem_open"|"runtime"|null, points_decay_factor: float|null, score_mode: "all_or_nothing"|"max_per_group"|"partial"|null, scoreboard: float|null, show_scoreboard_after: bool|null, start_time: \OmegaUp\Timestamp, submissions_gap: int|null, title: null|string, window_length?: int|null} $params
      */
     public function __construct($params) {
+        $languages = !is_null(
+            $params['languages']
+        ) ? explode(
+            ',',
+            $params['languages']
+        ) : null;
+        if (!is_null($languages)) {
+            \OmegaUp\Validators::validateValidSubset(
+                $languages,
+                'languages',
+                array_keys(\OmegaUp\Controllers\Run::SUPPORTED_LANGUAGES)
+            );
+        }
+
         $this->contestAlias = $params['alias'];
         $this->title = $params['title'];
         $this->description = $params['description'];
         $this->startTime = $params['start_time'];
         $this->finishTime = $params['finish_time'];
-        $this->lastUpadted = $params['last_updated'];
         $this->windowLength = $params['window_length'] ?? null;
-        $this->rerunId = $params['rerun_id'] ?? null;
         $this->admissionMode = $params['admission_mode'];
         $this->scoreboard = $params['scoreboard'];
         $this->pointsDecayFactor = $params['points_decay_factor'];
@@ -262,68 +239,11 @@ class ContestParams {
         $this->penaltyType = $params['penalty_type'];
         $this->penaltyCalcPolicy = $params['penalty_calc_policy'];
         $this->showScoreboardAfter = $params['show_scoreboard_after'];
-        $this->languages = $params['languages'] ?? null;
-        $this->urgent = $params['urgent'];
-        $this->recommended = $params['recommended'];
-        $this->archived = $params['archived'];
-        $this->certificateCutoff = $params['certificate_cutoff'] ?? null;
+        $this->languages = $languages;
         $this->certificatesStatus = $params['certificates_status'];
         $this->contestForTeams = $params['contest_for_teams'] ?? null;
         $this->defaultShowAllContestantsInScoreboard = $params['default_show_all_contestants_in_scoreboard'] ?? null;
         $this->scoreMode = $params['score_mode'];
-        $this->plagiarismTreshold = $params['plagiarism_treshold'];
         $this->checkPlagiarism = $params['check_plagiarism'];
-    }
-
-    /**
-     * Update properties of $object based on what is provided in this class.
-     *
-     * @param object $object
-     * @param array<int|string, string|array{transform?: callable(mixed):mixed, important?: bool, alias?: string}> $properties
-     * @return bool True if there were changes to any property marked as 'important'.
-     */
-    public function updateValueParams(
-        object $object,
-        array $properties
-    ): bool {
-        $importantChange = false;
-        foreach ($properties as $source => $info) {
-            /** @var null|callable(mixed):mixed */
-            $transform = null;
-            $important = false;
-            if (is_int($source)) {
-                $thisFieldName = $info;
-                $objectFieldName = $info;
-            } else {
-                $thisFieldName = $source;
-                if (isset($info['transform'])) {
-                    $transform = $info['transform'];
-                }
-                if (isset($info['important']) && $info['important'] === true) {
-                    $important = $info['important'];
-                }
-                if (!empty($info['alias'])) {
-                    $objectFieldName = $info['alias'];
-                } else {
-                    $objectFieldName = $thisFieldName;
-                }
-            }
-            // Get or calculate new value.
-            /** @var null|mixed */
-            $value = $this->$thisFieldName;
-            if (is_null($value)) {
-                continue;
-            }
-            if (!is_null($transform)) {
-                /** @var mixed */
-                $value = $transform($value);
-            }
-            // Important property, so check if it changes.
-            if ($important && !$importantChange) {
-                $importantChange = ($value != $object->$objectFieldName);
-            }
-            $object->$objectFieldName = $value;
-        }
-        return $importantChange;
     }
 }
